refactor: simplify CSS rules
Remove multiple flex containers. These are not needed to achieve the
desired layout. Instead we can achieve the same result by customizing
the normal flow with a couple of properties and by using only one flex
container.

- "text-align: center;" applied to the body element will be inherited by
every block container, thus horizontally aligning their content (the
title, images and text) to the center.
- since we don't use a flex container anymore, to achieve the spacing
between the title and the rest of the content, we replace the gap
property with margin-bottom.
- we use a single container to align and space the items. I also added
"flex-wrap: wrap;" to prevent clipping the images and text, when
resizing the browser to the minimum width.
- the items don't need to be flex containers, since the text is inside a
div element, which, because it is a block element, will already be
displayed on a new line, under the image.
- to ensure that changing the flex-basis in the future won't make our
items larger, we use max-width instead of width.